requirements.txt
Django>=4.1.6,<4.2
djangorestframework>=3.14.0,<3.15
psycopg-binary>=3.1.8,<3.2
drf-spectacular>=0.25.1,<0.26